requirements.txt
numpy
pandas
scikit-learn
tensorflow
matplotlib
searborn
opencv-python
fastapi
uvicorn
python-multipart
pillow
imutils